requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
django-redis==4.12.1
django-taggit==1.3.0
djangorestframework==3.11.0
drf-extensions==0.6.0
Faker==4.1.0
importlib-metadata==1.6.0
Markdown==3.2.2
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
zipp==3.1.0
